requirements.txt
aniso8601==7.0.0
anytree==2.8.0
apparmor==3.0.1
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.2.10
asn1crypto==1.4.0
bauh==0.9.16
bcrypt==3.2.0
Beaker==1.11.0
beautifulsoup4==4.9.3
Brlapi==0.8.2
btrfsutil==5.12
CacheControl==0.12.6
cachetools==4.2.2
caffeine-ng==3.5.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
Cheetah3==3.2.6.post1
cheroot==8.5.2
CherryPy==18.6.0
chrome-gnome-shell==0.0.0
cloudinary==1.25.0
colorama==0.4.4
coloredlogs==15.0
colorlog==5.0.1
configobj==5.1.0.dev0
contextlib2==0.6.0.post1
cryptography==3.4.7
cupshelpers==1.0
decorator==5.0.7
defusedxml==0.7.1
discid==1.2.0
distlib==0.3.1
distro==1.5.0
dj-database-url==0.5.0
dj-rest-auth==2.1.7
Django==3.1.2
django-allauth==0.44.0
django-cors-headers==3.6.0
django-extensions==3.1.3
django-heroku==0.3.1
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.1
dnspython==1.16.0
docopt==0.6.2
duplicity==0.8.18
ewmh==0.1.6
fasteners==0.15
feedparser==5.2.1
filelock==3.0.12
future==0.18.2
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gpg==1.15.1
graphene==2.1.8
graphene-django==2.14.0
graphql-core==2.3.2
graphql-relay==2.0.1
gufw==21.4.0
gunicorn==20.0.4
html5lib==1.1
httplib2==0.19.1
humanfriendly==9.1
idna==3.1
img2pdf==0.4.0
importlib-metadata==4.0.1
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.itertools==6.0.1
jaraco.logging==3.1.0
jaraco.stream==3.0.2
jaraco.text==3.5.0
layoutswitcherlib==0.8.17
lensfun==0.3.95
LibAppArmor==3.0.1
libfdt==1.6.0
louis==3.17.0
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
meson==0.58.0
mistune==0.8.4
mock==3.0.5
monotonic==1.6
Mopidy-ALSAMixer==2.0.1
Mopidy-Iris==3.56.1
Mopidy-Local==3.2.1
Mopidy-MPD==3.1.0
Mopidy-MPRIS==3.0.3
Mopidy-Spotify-Web==0.3.0
more-itertools==8.7.0
msgpack==1.0.2
mutagen==1.45.1
netsnmp-python==1.0a1
notify2==0.3.1
npyscreen==4.10.5
numpy==1.20.2
oauth2client==4.1.3
oauthlib==3.1.1
ocrmypdf==12.0.3
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.21.2
paramiko==2.7.2
pbr==5.6.0
pdfminer.six==20201018
pep517==0.10.0
picard==2.6.2
pikepdf==2.12.0
Pillow==8.2.0
pipenv==2021.5.29
pluggy==0.13.1
ply==3.11
portend==2.7.1
praw==7.2.0
prawcore==2.0.0
progress==1.5
promise==2.3
protobuf==3.16.0
psutil==5.8.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pulsectl==21.5.6
pwquality==1.4.4
pyalsaaudio==0.9.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycparser==2.20
pycryptodomex==3.9.9
pycups==2.0.1
pycurl==7.43.0.6
pydbus==0.6.0
PyDrive==1.3.1
Pygments==2.9.0
PyGObject==3.40.1
pyinotify==0.9.6
PyJWT==2.1.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt4-sip==4.19.25
PyQt5==5.15.4
PyQt5-sip==12.8.1
pysmbc==1.0.23
PySocks==1.7.1
python-dateutil==2.8.1
python-Levenshtein==0.12.2
python-xlib==0.29
python3-openid==3.2.0
pytz==2021.1
pyudev==0.22.0.dev20201112
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
resolvelib==0.5.4
retrying==1.3.3
rpm==4.16.1.3
rsa==4.7.2
Rx==1.6.1
sabyenc3==4.0.2
setproctitle==1.2.2
singledispatch==3.6.2
six==1.16.0
solaar==1.0.6
sortedcontainers==2.3.0
soupsieve==2.2.1
spotipy==2.16.1
sqlparse==0.4.1
team==1.0
tempora==4.0.2
text-unidecode==1.3
toml==0.10.2
torbrowser-launcher==0.3.3
tqdm==4.60.0
ufw==0.36
ulauncher==5.9.0
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.26.4
validate==5.1.0.dev0
variety==0.8.5
virtualenv==20.4.7
virtualenv-clone==0.5.4
webencodings==0.5.1
websocket-client==0.59.0
whitenoise==5.2.0
youtube-dl==2021.4.26
zc.lockfile==2.0
zipp==3.4.1
